packages/bedroomtv/pyproject.toml
[tool.poetry]
name = "bedroomtv"
version = "0.0.1"
description = ""
authors = ["Mariusz Kabala <mariusz@kabala.waw.pl>"]

[tool.poetry.dependencies]
python = "^3.7"
paho-mqtt = "^1.5.0"
python-json-logger = "^0.1.11"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
